requirements.txt
yfinance>=0.2.30
redis>=5.0
pandas>=2.1
ta>=0.11.0
python-telegram-bot>=21.0
apscheduler>=3.10
pyyaml>=6.0
watchdog>=3.0
httpx>=0.25
python-dotenv>=1.0
